Fix PulseAudio issue #5
Fixed some locking problems.
Removed circular references so that all objects are correctly
garbage collected.

+++ pyglet/media/drivers/pulse/adaptation.py
@@ -34,16 +34,20 @@
 from __future__ import print_function
 from __future__ import absolute_import
 
+import weakref
+
 from pyglet.media.drivers.base import AbstractAudioDriver, AbstractAudioPlayer
 from pyglet.media.events import MediaEvent
 from pyglet.media.exceptions import MediaException
 from pyglet.media.listener import AbstractListener
+from pyglet.debug import debug_print
 
 from . import lib_pulseaudio as pa
 from .interface import PulseAudioContext, PulseAudioContext, PulseAudioMainLoop, PulseAudioStream
 
 import pyglet
-_debug = pyglet.options['debug_media']
+
+_debug = debug_print('debug_media')
 
 
 class PulseAudioDriver(AbstractAudioDriver):
@@ -55,6 +59,9 @@ class PulseAudioDriver(AbstractAudioDriver):
 
         self._players = pyglet.app.WeakSet()
         self._listener = PulseAudioListener(self)
+
+    def __del__(self):
+        self.delete()
 
     def create_audio_player(self, playlist, player):
         assert self.context is not None
@@ -88,8 +95,6 @@ class PulseAudioDriver(AbstractAudioDriver):
     def delete(self):
         """Completely shut down pulseaudio client."""
         if self.mainloop is not None:
-            for player in self._players:
-                player.delete()
 
             with self.mainloop:
                 if self.context is not None:
@@ -107,7 +112,7 @@ class PulseAudioDriver(AbstractAudioDriver):
 
 class PulseAudioListener(AbstractListener):
     def __init__(self, driver):
-        self.driver = driver
+        self.driver = weakref.proxy(driver)
 
     def _set_volume(self, volume):
         self._volume = volume
@@ -148,14 +153,13 @@ class PulseAudioPlayer(AbstractAudioPlayer):
         audio_format = playlist.audio_format
         assert audio_format
 
-        with self.context.mainloop:
+        with driver.mainloop:
             self.stream = self.context.create_stream(audio_format)
             self.stream.push_handlers(self)
             self.stream.connect_playback()
             assert self.stream.is_ready
 
-        if _debug:
-            print('PulseAudioPlayer: __init__ finished')
+        assert _debug('PulseAudioPlayer: __init__ finished')
 
     def on_write_needed(self, nbytes, underflow):
         if underflow:
@@ -169,8 +173,7 @@ class PulseAudioPlayer(AbstractAudioPlayer):
                 self._time_sync_operation.delete()
                 self._time_sync_operation = None
             if self._time_sync_operation is None:
-                if _debug:
-                    print('PulseAudioPlayer: trigger timing info update')
+                assert _debug('PulseAudioPlayer: trigger timing info update')
                 self._time_sync_operation = self.stream.update_timing_info(self._process_events)
 
     def seek(self, timestamp):
@@ -178,7 +181,7 @@ class PulseAudioPlayer(AbstractAudioPlayer):
         self.audio_diff_cum = 0.0
         nbytes = 1 * self.playlist.audio_format.bytes_per_second
         while True:
-            audio_data = self.get_audio_data()
+            audio_data = self._get_audio_data()
             assert _debug("Seeking audio timestamp {:.2f} sec.".format(timestamp))
             
             if (audio_data is None or
@@ -188,10 +191,10 @@ class PulseAudioPlayer(AbstractAudioPlayer):
             assert _debug("Got audio packet starting at {:.2f} sec. "
                            "Skipping it.".format(audio_data.timestamp))
 
-            self._audiodata_buffer = None
+            self._current_audio_data = None
 
         if audio_data is not None:
-            assert _debug_media('Writing {} bytes'.format(audio_data.length))
+            assert _debug('Writing {} bytes'.format(audio_data.length))
             self._write_to_stream(audio_data.length)
 
     def _get_audio_data(self, nbytes=None):
@@ -202,16 +205,15 @@ class PulseAudioPlayer(AbstractAudioPlayer):
                 nbytes = min_bytes
             else:
                 nbytes = min(min_bytes, nbytes)
-            if _debug:
-                print('PulseAudioPlayer: Try to get {} bytes of audio data'.format(nbytes))
+            assert _debug('PulseAudioPlayer: Try to get {} bytes of audio data'.format(nbytes))
             compensation_time = self.get_audio_time_diff()
             self._current_audio_data = self.playlist.get_audio_data(nbytes, compensation_time)
             self._schedule_events()
-        if _debug:
-            if self._current_audio_data is None:
-                print('PulseAudioPlayer: No audio data available')
-            else:
-                print('PulseAudioPlayer: Got {} bytes of audio data'.format(self._current_audio_data.length))
+        if self._current_audio_data is None:
+            assert _debug('PulseAudioPlayer: No audio data available')
+        else:
+            assert _debug('PulseAudioPlayer: Got {} bytes of audio data'.format(
+                           self._current_audio_data.length))
         return self._current_audio_data
 
     def _has_audio_data(self):
@@ -229,22 +231,19 @@ class PulseAudioPlayer(AbstractAudioPlayer):
             for event in self._current_audio_data.events:
                 event_index = self._write_index + event.timestamp * \
                     self.playlist.audio_format.bytes_per_second
-                if _debug:
-                    print('PulseAudioPlayer: Schedule event at index {}'.format(event_index))
+                assert _debug('PulseAudioPlayer: Schedule event at index {}'.format(event_index))
                 self._events.append((event_index, event))
 
     def _write_to_stream(self, nbytes=None):
         if nbytes is None:
             nbytes = self.stream.writable_size
-        if _debug:
-            print('PulseAudioPlayer: Requested to write %d bytes to stream' % nbytes)
+        assert _debug('PulseAudioPlayer: Requested to write %d bytes to stream' % nbytes)
 
         seek_mode = pa.PA_SEEK_RELATIVE
         if self._clear_write:
             seek_mode = pa.PA_SEEK_RELATIVE_ON_READ
             self._clear_write = False
-            if _debug:
-                print('PulseAudioPlayer: Clear buffer')
+            assert _debug('PulseAudioPlayer: Clear buffer')
 
         while self._has_audio_data() and nbytes > 0:
             audio_data = self._get_audio_data()
@@ -257,8 +256,8 @@ class PulseAudioPlayer(AbstractAudioPlayer):
             self._timestamps.append((self._write_index, audio_data.timestamp))
             self._write_index += consumption
 
-            if _debug:
-                print('PulseAudioPlayer: Actually wrote %d bytes to stream' % consumption)
+            assert _debug('PulseAudioPlayer: Actually wrote {} bytes '
+                          'to stream'.format(consumption))
             self._consume_audio_data(consumption)
 
             nbytes -= consumption
@@ -272,67 +271,55 @@ class PulseAudioPlayer(AbstractAudioPlayer):
                 op.delete()  # Explicit delete to prevent locking
 
     def _handle_underflow(self):
-        if _debug:
-            print('Player: underflow')
+        assert _debug('Player: underflow')
         if self._has_audio_data():
             self._write_to_stream()
         else:
             self._add_event_at_write_index('on_eos')
 
     def _process_events(self):
-        if _debug:
-            print('PulseAudioPlayer: Process events')
+        assert _debug('PulseAudioPlayer: Process events')
         if not self._events:
-            if _debug:
-                print('PulseAudioPlayer: No events')
+            assert _debug('PulseAudioPlayer: No events')
             return
 
         # Assume this is called after time sync
         timing_info = self.stream.get_timing_info()
         if not timing_info:
-            if _debug:
-                print('PulseAudioPlayer: No timing info to process events')
+            assert _debug('PulseAudioPlayer: No timing info to process events')
             return
 
         read_index = timing_info.read_index
-        if _debug:
-            print('PulseAudioPlayer: Dispatch events at index {}'.format(read_index))
+        assert _debug('PulseAudioPlayer: Dispatch events at index {}'.format(read_index))
 
         while self._events and self._events[0][0] <= read_index:
             _, event = self._events.pop(0)
-            if _debug:
-                print('PulseAudioPlayer: Dispatch event', event)
+            assert _debug('PulseAudioPlayer: Dispatch event', event)
             event._sync_dispatch_to_player(self.player)
 
     def _add_event_at_write_index(self, event_name):
-        if _debug:
-            print('PulseAudioPlayer: Add event at index {}'.format(self._write_index))
+        assert _debug('PulseAudioPlayer: Add event at index {}'.format(self._write_index))
         self._events.append((self._write_index, MediaEvent(0., event_name)))
 
-
-    def __del__(self):
-        try:
-            self.delete()
-        except:
-            pass
-
     def delete(self):
-        if _debug:
-            print('PulseAudioPlayer.delete')
+        assert _debug('Delete PulseAudioPlayer')
+
+        self.stream.pop_handlers()
+        if self.driver.mainloop is None:
+            assert _debug('PulseAudioDriver already deleted. '
+                      'PulseAudioPlayer could not clean up properly.')
+            return
 
         if self._time_sync_operation is not None:
             with self._time_sync_operation:
                 self._time_sync_operation.delete()
             self._time_sync_operation = None
 
-        if self.stream is not None:
-            with self.stream:
-                self.stream.delete()
-            self.stream = None
+        self.stream.delete()
+        self.stream = None
 
     def clear(self):
-        if _debug:
-            print('PulseAudioPlayer.clear')
+        assert _debug('PulseAudioPlayer.clear')
         super(PulseAudioPlayer, self).clear()
         self._clear_write = True
         self._write_index = self._get_read_index()
@@ -344,27 +331,23 @@ class PulseAudioPlayer(AbstractAudioPlayer):
             self.stream.prebuf().wait()
 
     def play(self):
-        if _debug:
-            print('PulseAudioPlayer.play')
+        assert _debug('PulseAudioPlayer.play')
 
         with self.stream:
             if self.stream.is_corked:
                 self.stream.resume().wait().delete()
-                if _debug:
-                    print('PulseAudioPlayer: Resumed playback')
+                assert _debug('PulseAudioPlayer: Resumed playback')
             if self.stream.underflow:
                 self._write_to_stream()
             if not self._has_audio_data():
                 self.stream.trigger().wait().delete()
-                if _debug:
-                    print('PulseAudioPlayer: Triggered stream for immediate playback')
+                assert _debug('PulseAudioPlayer: Triggered stream for immediate playback')
             assert not self.stream.is_corked
 
         self._playing = True
 
     def stop(self):
-        if _debug:
-            print('PulseAudioPlayer.stop')
+        assert _debug('PulseAudioPlayer.stop')
 
         with self.stream:
             if not self.stream.is_corked:
@@ -382,8 +365,7 @@ class PulseAudioPlayer(AbstractAudioPlayer):
         else:
             read_index = 0
 
-        if _debug:
-            print('_get_read_index ->', read_index)
+        assert _debug('_get_read_index ->', read_index)
         return read_index
 
     def _get_write_index(self):
@@ -393,14 +375,12 @@ class PulseAudioPlayer(AbstractAudioPlayer):
         else:
             write_index = 0
 
-        if _debug:
-            print('_get_write_index ->', write_index)
+        assert _debug('_get_write_index ->', write_index)
         return write_index
 
     def get_time(self):
         if not self._read_index_valid:
-            if _debug:
-                print('get_time <_read_index_valid = False> -> None')
+            assert _debug('get_time <_read_index_valid = False> -> None')
             return
 
         read_index = self._get_read_index()
@@ -420,8 +400,7 @@ class PulseAudioPlayer(AbstractAudioPlayer):
         bytes_per_second = self.playlist.audio_format.bytes_per_second
         time = timestamp + (read_index - write_index) / float(bytes_per_second)
 
-        if _debug:
-            print('get_time ->', time)
+        assert _debug('get_time ->', time)
         return time
 
     def set_volume(self, volume):
